+++ akeneo-design-system/src/components/Button/Button.tsx
@@ -1,7 +1,8 @@
-import React, {ReactNode, Ref, SyntheticEvent} from 'react';
+import React, {isValidElement, ReactNode, Ref, SyntheticEvent} from 'react';
 import styled, {css} from 'styled-components';
-import {AkeneoThemedProps, getColorForLevel, getFontSize, Level} from '../../theme';
+import {AkeneoThemedProps, getColor, getColorForLevel, getFontSize, Level} from '../../theme';
 import {Override} from '../../shared';
+import {IconProps} from '../../icons';
 
 type ButtonSize = 'small' | 'default';
 
@@ -69,7 +70,7 @@ const getColorStyle = ({
   if (ghost) {
     return css`
       color: ${getColorForLevel(level, disabled ? 80 : 120)};
-      background-color: white;
+      background-color: ${getColor('white')};
       border-color: ${getColorForLevel(level, disabled ? 60 : 100)};
 
       &:hover:not([disabled]) {
@@ -86,7 +87,7 @@ const getColorStyle = ({
   }
 
   return css`
-    color: white;
+    color: ${getColor('white')};
     background-color: ${getColorForLevel(level, disabled ? 40 : 100)};
 
     &:hover:not([disabled]) {
@@ -99,7 +100,7 @@ const getColorStyle = ({
   `;
 };
 
-const ContainerStyle = css<
+const Container = styled.button<
   {
     level: Level;
     ghost: boolean;
@@ -107,7 +108,9 @@ const ContainerStyle = css<
     size: ButtonSize;
   } & AkeneoThemedProps
 >`
-  display: inline-block;
+  display: inline-flex;
+  align-items: center;
+  gap: 10px;
   border-width: 1px;
   font-size: ${getFontSize('default')};
   font-weight: 400;
@@ -116,22 +119,17 @@ const ContainerStyle = css<
   border-style: ${({ghost}) => (ghost ? 'solid' : 'none')};
   padding: ${({size}) => (size === 'small' ? '0 10px' : '0 15px')};
   height: ${({size}) => (size === 'small' ? '24px' : '32px')};
-  line-height: ${({size}) => (size === 'small' ? 24 : 32) - 2}px;
-  outline-color: ${({level}) => getColorForLevel(level, 100)};
   cursor: ${({disabled}) => (disabled ? 'not-allowed' : 'pointer')};
   font-family: inherit;
   transition: background-color 0.1s ease;
+  outline-style: none;
+  text-decoration: none;
+
+  &:focus {
+    box-shadow: 0 0 0 2px ${getColor('blue', 40)};
+  }
 
   ${getColorStyle}
-`;
-
-const ButtonContainer = styled.button`
-  ${ContainerStyle}
-`;
-
-const LinkContainer = styled.a`
-  text-decoration: none;
-  ${ContainerStyle}
 `;
 
 /**
@@ -156,20 +154,15 @@ const Button = React.forwardRef<HTMLButtonElement, ButtonProps>(
     }: ButtonProps,
     forwardedRef: Ref<HTMLButtonElement>
   ) => {
-    if (undefined !== href && undefined !== onClick) {
-      throw new Error('Button cannot have both `href` and `onClick` props');
-    }
-
     const handleAction = (event: SyntheticEvent) => {
       if (disabled || undefined === onClick) return;
 
       onClick(event);
     };
 
-    const Component = undefined !== href ? LinkContainer : ButtonContainer;
-
     return (
-      <Component
+      <Container
+        as={undefined !== href ? 'a' : 'button'}
         level={level}
         ghost={ghost}
         disabled={disabled}
@@ -185,8 +178,14 @@ const Button = React.forwardRef<HTMLButtonElement, ButtonProps>(
         href={disabled ? undefined : href}
         {...rest}
       >
-        {children}
-      </Component>
+        {React.Children.map(children, child => {
+          if (isValidElement<IconProps>(child)) {
+            return React.cloneElement(child, {size: child.props.size ?? 18});
+          }
+
+          return child;
+        })}
+      </Container>
     );
   }
 );