+++ akeneo-design-system/src/components/Button/Button.unit.tsx
@@ -66,20 +66,6 @@ test('it displays an anchor when providing a `href`', () => {
   expect(screen.getByText('Hello').closest('a')).toHaveAttribute('href', 'https://akeneo.com/');
 });
 
-test('it throws when trying to pass href and onClick', () => {
-  const mockConsole = jest.spyOn(console, 'error').mockImplementation();
-
-  expect(() => {
-    render(
-      <Button onClick={jest.fn()} href="#nice">
-        Hello
-      </Button>
-    );
-  }).toThrowError();
-
-  mockConsole.mockRestore();
-});
-
 test('it does not trigger onClick when disabled', () => {
   const onClick = jest.fn();
   render(
@@ -111,14 +97,16 @@ test('Button supports forwardRef', () => {
       My button
     </Button>
   );
+
   expect(ref.current).not.toBe(null);
 });
 
 test('Button supports ...rest props', () => {
-  const {container} = render(
-    <Button onClick={jest.fn()} data-my-attribute="my_value">
+  render(
+    <Button onClick={jest.fn()} data-testid="my_value">
       My button
     </Button>
   );
-  expect(container.querySelector('[data-my-attribute="my_value"]')).toBeInTheDocument();
+
+  expect(screen.getByTestId('my_value')).toBeInTheDocument();
 });