+++ akeneo-design-system/src/hooks/index.ts
@@ -1,5 +1,7 @@
 export * from './useAutoFocus';
 export * from './useId';
+export * from './useProgress';
+export * from './useSelection';
 export * from './useShortcut';
 export * from './useTheme';
 export * from './useBooleanState';