+++ src/Akeneo/Connectivity/Connection/back/tests/Integration/Persistence/Dbal/Query/SelectEventsApiRequestCountWithinLastHourQueryIntegration.php
@@ -3,7 +3,7 @@ declare(strict_types=1);
 
 namespace Akeneo\Connectivity\Connection\back\tests\Integration\Persistence\Dbal\Query;
 
-use Akeneo\Connectivity\Connection\back\tests\Integration\Fixtures\EventsApiRequestCountLoader;
+use Akeneo\Connectivity\Connection\Tests\CatalogBuilder\EventsApiRequestCountLoader;
 use Akeneo\Connectivity\Connection\Domain\Audit\Persistence\Query\SelectEventsApiRequestCountWithinLastHourQuery;
 use Akeneo\Test\Integration\TestCase;
 use PHPUnit\Framework\Assert;