+++ src/Akeneo/Pim/Enrichment/Component/Product/Connector/Processor/Normalization/ProductProcessor.php
@@ -3,7 +3,9 @@
 namespace Akeneo\Pim\Enrichment\Component\Product\Connector\Processor\Normalization;
 
 use Akeneo\Pim\Enrichment\Component\Product\Connector\Processor\FilterValues;
+use Akeneo\Pim\Enrichment\Component\Product\Connector\UseCase\GetProductsWithQualityScoresInterface;
 use Akeneo\Pim\Enrichment\Component\Product\Model\EntityWithFamilyInterface;
+use Akeneo\Pim\Enrichment\Component\Product\Model\ProductInterface;
 use Akeneo\Pim\Enrichment\Component\Product\Model\ProductModelInterface;
 use Akeneo\Pim\Enrichment\Component\Product\ValuesFiller\FillMissingValuesInterface;
 use Akeneo\Pim\Structure\Component\Repository\AttributeRepositoryInterface;
@@ -44,18 +46,22 @@ class ProductProcessor implements ItemProcessorInterface, StepExecutionAwareInte
     /** @var FillMissingValuesInterface */
     protected $fillMissingProductModelValues;
 
+    private ?GetProductsWithQualityScoresInterface $getProductsWithQualityScores;
+
     public function __construct(
         NormalizerInterface $normalizer,
         IdentifiableObjectRepositoryInterface $channelRepository,
         AttributeRepositoryInterface $attributeRepository,
         BulkMediaFetcher $mediaFetcher,
-        FillMissingValuesInterface $fillMissingProductModelValues
+        FillMissingValuesInterface $fillMissingProductModelValues,
+        ?GetProductsWithQualityScoresInterface $getProductsWithQualityScores = null
     ) {
         $this->normalizer          = $normalizer;
         $this->channelRepository   = $channelRepository;
         $this->attributeRepository = $attributeRepository;
         $this->mediaFetcher        = $mediaFetcher;
         $this->fillMissingProductModelValues = $fillMissingProductModelValues;
+        $this->getProductsWithQualityScores = $getProductsWithQualityScores;
     }
 
     /**
@@ -98,6 +104,15 @@ class ProductProcessor implements ItemProcessorInterface, StepExecutionAwareInte
                     return !in_array($attributeCode, $mediaAttributes);
                 },
                 ARRAY_FILTER_USE_KEY
+            );
+        }
+
+        if (null !== $this->getProductsWithQualityScores && $product instanceof ProductInterface && $this->hasFilterOnQualityScore($parameters)) {
+            $productStandard = $this->getProductsWithQualityScores->fromNormalizedProduct(
+                $product->getIdentifier(),
+                $productStandard,
+                $structure['scope'] ?? null,
+                $structure['locales'] ?? []
             );
         }
 
@@ -170,5 +185,17 @@ class ProductProcessor implements ItemProcessorInterface, StepExecutionAwareInte
     {
         return isset($parameters->get('filters')['structure']['attributes'])
             && !empty($parameters->get('filters')['structure']['attributes']);
+    }
+
+    private function hasFilterOnQualityScore(JobParameters $parameters): bool
+    {
+        foreach ($parameters->get('filters')['data'] ?? [] as $filter) {
+            $field = $filter['field'] ?? null;
+            if ($field === 'quality_score_multi_locales') {
+                return true;
+            }
+        }
+
+        return false;
     }
 }