+++ src/Akeneo/Platform/Bundle/UIBundle/Resources/public/js/form/common/main-image.js
@@ -6,11 +6,12 @@
  * @copyright 2017 Akeneo SAS (http://www.akeneo.com)
  * @license   http://opensource.org/licenses/osl-3.0.php  Open Software License (OSL 3.0)
  */
-define(['underscore', 'pim/form', 'pim/template/form/main-image', 'pim/media-url-generator'], function (
+define(['underscore', 'pim/form', 'pim/template/form/main-image', 'pim/media-url-generator', 'jquery'], function (
   _,
   BaseForm,
   template,
-  MediaUrlGenerator
+  MediaUrlGenerator,
+  $
 ) {
   return BaseForm.extend({
     tagName: 'img',
@@ -22,6 +23,7 @@ define(['underscore', 'pim/form', 'pim/template/form/main-image', 'pim/media-url
      */
     initialize: function (config) {
       this.config = config.config;
+      this.imagePath = null;
 
       BaseForm.prototype.initialize.apply(this, arguments);
     },
@@ -43,6 +45,10 @@ define(['underscore', 'pim/form', 'pim/template/form/main-image', 'pim/media-url
 
       this.el.src = this.getPath();
 
+      this.$el.one('error', function () {
+        $(this).attr('src', MediaUrlGenerator.getMediaShowUrl(null, 'thumbnail_small'));
+      });
+
       return BaseForm.prototype.render.apply(this, arguments);
     },
 
@@ -56,7 +62,7 @@ define(['underscore', 'pim/form', 'pim/template/form/main-image', 'pim/media-url
         return this.config.path;
       }
 
-      const filePath = _.result(this.getFormData().meta.image, 'filePath', null);
+      const filePath = this.imagePath ?? _.result(this.getFormData().meta.image, 'filePath', null);
 
       if (filePath === null && undefined !== this.config.fallbackPath) {
         return this.config.fallbackPath;